+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+* Only return HTML attributes with not-null value
+
 ## [0.2.2] - 2022-01-10
 
 * Trim combined classes